+++ src/main/java/org/bytedeco/javacpp/presets/javacpp.java
@@ -50,7 +50,13 @@ import org.bytedeco.javacpp.annotation.Properties;
                        "api-ms-win-core-sysinfo-l1-1-0", "api-ms-win-core-synch-l1-2-0", "api-ms-win-core-console-l1-1-0", "api-ms-win-core-debug-l1-1-0",
                        "api-ms-win-core-rtlsupport-l1-1-0", "api-ms-win-core-processthreads-l1-1-1", "api-ms-win-core-file-l1-2-0", "api-ms-win-core-profile-l1-1-0",
                        "api-ms-win-core-memory-l1-1-0", "api-ms-win-core-util-l1-1-0", "api-ms-win-core-interlocked-l1-1-0", "ucrtbase",
-                       "vcruntime140", "vcruntime140_1", "vcruntime140_threads", "msvcp140", "msvcp140_1", "msvcp140_2", "concrt140", "vcomp140", "libomp140.i386", "libomp140.x86_64"}
+                       "vcruntime140", "vcruntime140_1", "vcruntime140_threads", "msvcp140", "msvcp140_1", "msvcp140_2", "concrt140", "vcomp140", "libomp140.aarch64", "libomp140.i386", "libomp140.x86_64"}
+        ),
+        @Platform(
+            value = "windows-arm64",
+            preloadpath = {"C:/Program Files (x86)/Microsoft Visual Studio 14.0/VC/redist/arm64/Microsoft.VC140.CRT/",
+                           "C:/Program Files (x86)/Microsoft Visual Studio 14.0/VC/redist/arm64/Microsoft.VC140.OpenMP/",
+                           "C:/Program Files (x86)/Windows Kits/10/Redist/ucrt/DLLs/arm64/"}
         ),
         @Platform(
             value = "windows-x86",
@@ -77,6 +83,9 @@ public class javacpp implements LoadEnabled {
         String ucrtversion = System.getenv("UCRTVersion");
         if (ucrtsdkdir != null && ucrtsdkdir.length() > 0 && ucrtversion != null && ucrtversion.length() > 0) {
             switch (platform) {
+                case "windows-arm64":
+                    preloadpaths.add(0, ucrtsdkdir + "\\Redist\\" + ucrtversion + "\\ucrt\\DLLs\\arm64");
+                    break;
                 case "windows-x86":
                     preloadpaths.add(0, ucrtsdkdir + "\\Redist\\" + ucrtversion + "\\ucrt\\DLLs\\x86");
                     break;
@@ -91,6 +100,14 @@ public class javacpp implements LoadEnabled {
         String vcredistdir = System.getenv("VCToolsRedistDir");
         if (vcredistdir != null && vcredistdir.length() > 0) {
             switch (platform) {
+                case "windows-arm64":
+                    preloadpaths.add(0, vcredistdir + "\\arm64\\Microsoft.VC143.CRT");
+                    preloadpaths.add(1, vcredistdir + "\\arm64\\Microsoft.VC143.OpenMP");
+                    preloadpaths.add(2, vcredistdir + "\\arm64\\Microsoft.VC142.CRT");
+                    preloadpaths.add(3, vcredistdir + "\\arm64\\Microsoft.VC142.OpenMP");
+                    preloadpaths.add(4, vcredistdir + "\\arm64\\Microsoft.VC141.CRT");
+                    preloadpaths.add(5, vcredistdir + "\\arm64\\Microsoft.VC141.OpenMP");
+                    break;
                 case "windows-x86":
                     preloadpaths.add(0, vcredistdir + "\\x86\\Microsoft.VC143.CRT");
                     preloadpaths.add(1, vcredistdir + "\\x86\\Microsoft.VC143.OpenMP");